some more unsigned->size_t fixes

+++ src/libFLAC/bitbuffer.c
@@ -580,7 +580,7 @@ unsigned FLAC__bitbuffer_get_input_bytes_unconsumed(const FLAC__BitBuffer *bb)
 	return (bb->total_bits - bb->total_consumed_bits) >> 3;
 }
 
-void FLAC__bitbuffer_get_buffer(FLAC__BitBuffer *bb, const FLAC__byte **buffer, unsigned *bytes)
+void FLAC__bitbuffer_get_buffer(FLAC__BitBuffer *bb, const FLAC__byte **buffer, size_t *bytes)
 {
 	FLAC__ASSERT((bb->consumed_bits & 7) == 0 && (bb->bits & 7) == 0);
 #if FLAC__BITS_PER_BLURB == 8
